Extract output section into its own component
