+++ client/App.tsx
@@ -6,21 +6,13 @@ import ExpressStatic from './components/ExpressStatic';
 import RouteState from './types/RouteState';
 import routeReducer from './reducers/RouteReducer';
 import { RouteContext, RouteDispatchContext } from './RouteContext';
+import Output from './components/Output';
 
 const globalStyle = css({
   fontFamily: 'Roboto, Helvetica, Arial, sans-serif',
   fontWeight: 400,
   fontSize: '1rem',
   color: 'rgba(0, 0, 0, 0.87)'
-});
-
-const serverSetupStyle = css({
-  margin: '2rem 1rem',
-  display: 'flex',
-  '& p, pre': {
-    whiteSpace: 'nowrap',
-    margin: 'auto 0.5rem'
-  }
 });
 
 const initialRouteProps: RouteState = {
@@ -41,9 +33,7 @@ const App = (): JSX.Element => {
           <hr />
           <ExpressStatic />
           <hr />
-          <section id='result' css={serverSetupStyle}>
-            <p>-- would be served at <code>{`${routeProps.hostname}:${routeProps.port}/${routeProps.httpRoute}`}cat_meme.gif</code>.</p>
-          </section>
+          <Output />
         </RouteDispatchContext.Provider>
       </RouteContext.Provider>
     </main>
